Moved Ip tests to a directory
